Extracting out port allocation
