+++ Source/DevCentral/main/workspaces/for_Workspaces/when_constructing_and_root_does_not_exist.ts
@@ -15,6 +15,7 @@ describe('when constructing and root does not exist', () => {
         dependencies.fileSystem,
         dependencies.applicationLoader,
         dependencies.microserviceLoader,
+        dependencies.microservicePortsAllocator,
         dependencies.logger
     );
